useredit

+++ app/Http/Controllers/UsersController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\User;
+use Carbon\Carbon;
 class UsersController extends Controller
 {
     //public function userdelete($id)
@@ -19,10 +20,14 @@ class UsersController extends Controller
     
     public function destroy($id)
     {
-       $user = User::findOrFail($id);
-       \Auth::logout();
-       $user->delete();
-       return redirect("/");
+        $user = User::findOrFail($id);
+        if (\Auth::id() === $user->id){
+           \Auth::logout();
+           $user->delete();
+        }
+            //session()->flash('flash_message', '退部しました。今までお世話になりました。');
+            return redirect("/");
+       
     }
     
     public function index()
@@ -47,4 +52,51 @@ class UsersController extends Controller
         ]);
     }
     
+    public function edit($id){
+        
+        $user = User::findOrFail($id);
+        $birth = $user->birthday;
+        $date = new Carbon($birth); 
+        $myyear = $date->year;
+        $mymonth = $date->month;
+        $myday = $date->day;
+        
+        if(\Auth::id() ===$user->id){
+            return view("users.edit" , [
+                "user" => $user,
+                "myyear"=>$myyear,
+                "mymonth"=>$mymonth,
+                "myday"=>$myday,
+            ]);
+        }
+    }
+    
+    public function update(Request $request, $id)
+    {
+        // $request->validate([ 
+        //     "content" => "required" ,
+        //     "status" => "required|max:10" ,
+        // ]);
+        
+        $user = User::findOrFail($id) ;
+        if (\Auth::id() === $user->id){
+            $user->name = $request->name ;
+            $user->birthday = $request->year."-" .sprintf('%02d',$request->month)."-".sprintf('%02d',$request->day);
+            $user->mynum = $request->mynum;
+            $user->dominant_def = $request->dominant_def;
+            $user->dominant_bat = $request->dominant_bat;
+            $user->save();
+        }
+        
+        return redirect('/');
+    }
+    public function userdelete($id)
+    {
+        $user = User::findOrFail($id);
+        
+            return view("users.delete" , [
+                    "user" => $user ,
+            ]);
+        
+    }
 }
